PLANET-6122 Improve EditableBackground
Add MediaUploadCheck and remove EditorControls component
